CLEAR-2: fixed services classes
