+++ order-service/src/main/java/com/example/orderservice/service/CategoryService.java
@@ -1,0 +1,7 @@
+package com.example.orderservice.service;
+
+import com.example.orderservice.entity.Category;
+
+public interface CategoryService {
+    Category findCategoryByName(String name);
+}
