+++ order-service/src/main/java/com/example/orderservice/service/impl/UserServiceImpl.java
@@ -1,0 +1,52 @@
+package com.example.orderservice.service.impl;
+
+import com.example.orderservice.entity.Role;
+import com.example.orderservice.entity.User;
+import com.example.orderservice.exception.EntityNotFoundException;
+import com.example.orderservice.repository.UserRepository;
+import com.example.orderservice.service.UserService;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+@Slf4j
+public class UserServiceImpl implements UserService {
+
+    private final UserRepository userRepository;
+
+    @Override
+    public User getById(Long id) throws EntityNotFoundException {
+        log.info("Getting user with id {}",id);
+
+        return userRepository.findById(id)
+                .orElseThrow(()-> new EntityNotFoundException("There is not user with id "+id));
+    }
+
+    @Override
+    public User getByUserEmail(String userEmail) throws EntityNotFoundException {
+        log.info("Getting user with email {}",userEmail);
+
+        return userRepository.findUserByEmail(userEmail)
+                .orElseThrow(() -> new EntityNotFoundException("There is not user with email "+userEmail));
+    }
+
+    @Override
+    public User getManager() throws EntityNotFoundException {
+        log.info("Getting random manager to assign ticket");
+
+        return userRepository.findUserByRole(Role.MANAGER)
+                .orElseThrow(()->new EntityNotFoundException("No managers"));
+    }
+
+    @Override
+    public User getEngineer() throws EntityNotFoundException {
+        log.info("Getting random engineer to approve ticket");
+
+        return userRepository.findUserByRole(Role.ENGINEER)
+                .orElseThrow(()->new EntityNotFoundException("No engineers"));
+    }
+}
